requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
sqlparse==0.5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1
openai==1.54.1
python-dotenv==1.0.1